// src/core/api/generated/spring/schemas/CustomerDTOStatus.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type CustomerDTOStatus = typeof CustomerDTOStatus[keyof typeof CustomerDTOStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CustomerDTOStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  ARCHIVED: 'ARCHIVED',
} as const;